more progress

+++ sock/rpc/rpc2/event.h
@@ -12,7 +12,7 @@ class remote_callback
 {
 public:
 	int socket;
-	int addr;
+	unsigned long addr;
 };
 
 class event
